[refactor] #133 Alert Live, Test 분리
